revised my files and created rspec to test
